servicio_login/config.py
import os
from dotenv import load_dotenv

load_dotenv()

class Config:
    SECRET_KEY = os.getenv('SECRET_KEY', 'default_secret_key')
    SQLALCHEMY_DATABASE_URI = os.getenv('DATABASE_URL')
    SQLALCHEMY_TRACK_MODIFICATION = False
    JWT_SECRET_KEY = os.get_getenv('JWT_SECRET_KEY', 'jwt_secret_key')