servicio_login/requirements.txt
Flask
Flask-JWT-Extended
Flask-SQLAlchemy
psycopg2
python-dotenv